+++ packages/apps/tools/src/components/Global/ModuleExplorer/SidePanel/outline.tsx
@@ -1,8 +1,10 @@
 import { contractParser } from '@kadena/pactjs-generator';
-import type { ITreeProps } from '@kadena/react-ui';
-import { Button, Heading, Text, Tree } from '@kadena/react-ui';
+import { Heading, Text, Tree } from '@kadena/react-ui';
 
 import type { IChainModule } from '../types';
+
+import type { Contract } from './utils';
+import { contractToTreeItems } from './utils';
 
 import useTranslation from 'next-translate/useTranslation';
 import React from 'react';
@@ -11,57 +13,6 @@ export interface IOutlineProps extends React.HTMLAttributes<HTMLDivElement> {
   selectedModule?: IChainModule;
   onInterfaceClick: (module: IChainModule) => void;
 }
-
-type Contract = ReturnType<typeof contractParser>[0][0]; // TODO: Should we improve this because it's a bit hacky?
-
-const contractToTreeItems = (
-  contract: Contract,
-  onInterfaceClick: IOutlineProps['onInterfaceClick'],
-  module: IOutlineProps['selectedModule'],
-): ITreeProps['items'] => {
-  const { usedInterface: interfaces, capabilities, functions } = contract;
-  const items: ITreeProps['items'] = [];
-
-  if (interfaces?.length) {
-    items.push({
-      title: 'Interfaces',
-      isOpen: true,
-      items: interfaces.map((i) => ({
-        title: (
-          <Button
-            onClick={() =>
-              onInterfaceClick({ chainId: module!.chainId, moduleName: i.name })
-            }
-            variant="compact"
-            icon="ExitToApp"
-          >
-            {i.name}
-          </Button>
-        ),
-      })),
-    });
-  }
-
-  if (capabilities?.length) {
-    items.push({
-      title: 'Capabilities',
-      items: capabilities.map((c) => ({
-        title: c.name,
-      })),
-    });
-  }
-
-  if (functions?.length) {
-    items.push({
-      title: 'Functions',
-      items: functions.map((f) => ({
-        title: f.name,
-      })),
-    });
-  }
-
-  return items;
-};
 
 const Outline = (props: IOutlineProps): React.JSX.Element => {
   const { selectedModule, onInterfaceClick } = props;
